requirements.txt
Flask==2.3.2
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.5
python-dotenv==1.0.0
requests==2.31.0
azure-core==1.33.0
azure-ai-inference==1.0.0b9
isodate==0.7.2
six==1.17.0
openai==1.63.0